Here is our testing script for micro motor setup
This is the code for using the small sparkfun micro motors and the
adafruit V2 motor shield.

This commit allowed for the discovery that the motors rev/min and wheel
size were hilariously too small.

This branch will likely terminate for the time being. This setup may be
valid if the motor/wheel setup is modified on the hardware end.