Dusting off this project & adding motor init to 0
Did some experimenting on the ESC of things that I have forgotten
regarding its behavior.

Important notes
	At this time is would seem 0 speed == 91 on the servo, and that
there is about a 15 swing for max speed in either direction.

	continuously writing to the sevo ultrarapidly will cause it to
"stall" or whatnot.

Was also playing with the ESC - not sure what the sound means on the
motors - but sometimes a motor will play twice. Maybe a side effect of
programming with everything plugged in? big doubt.

Between playing on the website, forgetting some arduino library syntax,
and needing to brush up on the ESC behavior - I'm calling it a close
after adding only the init motor function