This was odd... a test on the bidirectional ESC
I thought i had previously noted that the esc's don't work well
if you too rapidly write to them. So, I made a test here...

The test was that on startup, the motors were initialized to zero speed
(with many precautions trying to make sure that we didn't too rapidly
write to the ESC during the setup).

5 seconds after initializing the motors, the loop funciton ONLY writes
max speed to the motors.

The expecatation (based on what i thought I had previously seen with
this ESC) is that the motor would never start spinning due to too
rapidly writing... that didn't happen. This test branch is temporaily
closed due to getting happily unexpected results.