Here is the most basic code that should facilitate self balance
Here the 2 pwm's have been used to get the bi-directional ESC to drive
the motors. The Kp, Ki, and Kd are all for testing.

It was discovered in the testing from here though that the motors I had
selected for this method (brushless outrunners) cannot supply enough
torque. Upon applying the full weight of the device, the motors
practically stall.

I may be able to set up my PSU to allow for more amperage. I put a 5 amp
cap on the 5V supply and it didn't work. The ESC supports up to 20A, but
I don't want to use that much power, and don't think the torqe can be
overcome.