Making the first gen bracket for stepper motor.
I tried to add a tab that would hold the motor in the bracket

The tab does not work well with 3d print (it just snapped with no give.)

Also, after having printed it, the tolerance on the inside dimensions is too small.
I can't get the motor out because the pressure fit is too tight.